+++ src/vm/vm.hpp
@@ -30,8 +30,8 @@ class VM {
  public:
   std::vector<uint8_t> bytes;
   std::vector<Frame> frames;
-  std::unordered_map<int32_t, std::shared_ptr<TPV_ObjString>> str_table;
-  std::unordered_map<int32_t, std::shared_ptr<TPV_Obj>> heap;
+  std::unordered_map<size_t, std::shared_ptr<TPV_ObjString>> str_table;
+  std::unordered_map<size_t, std::shared_ptr<TPV_Obj>> heap;
   std::vector<Error> errors;
   FLAGS flags;
   uint32_t pc = 0;
